api 분리
