+++ src/users/user-provider.js
@@ -1,147 +1,11 @@
 const { User } = require('../schemas/user');
 const UserRefo = require('./user-repo');
 const jwtService = require('./jwt');
+const axios = require('axios');
+const qs = require('qs');
 require('dotenv').config();
 
-/* 예시
-  user-route.js 1, res.locals에 저장한 userInfo:::::: {
-    id: 2519073484,
-    connected_at: '2022-11-10T01:02:16Z',
-    properties: { nickname: '미뇽' },
-    kakao_account: {
-      profile_nickname_needs_agreement: false,
-      profile_image_needs_agreement: true,
-      profile: { nickname: '미뇽' },
-      has_email: true,
-      email_needs_agreement: false,
-      is_email_valid: true,
-      is_email_verified: true,
-      email: 'alsuddl25@naver.com'
-    }
-  }
-*/
-
 class UserProvider {
-  exUserGetToken = async (email) => {
-    // 존재하는 유저일 경우 토큰 발급하여 가져오기
-    const exUser = await User.findOne({
-      email: kakaoUserData.kakao_account.email,
-    });
-    console.log('여기는 user-provider.js 1, exUserGetToken, exUser:::', exUser);
-    if (exUser) {
-      const accessToken = await jwtService.createAccessToken(exUser._id);
-      console.log('여기는 user-provider.js 2, accessToken::::::', accessToken);
-      let spyWinRating, voteSpyRating, totalCount;
-      if (exUser.totalCount === 0) {
-        spyWinRating = 0;
-        voteSpyRating = 0;
-      } else if (exUser.spyPlayCount === 0) {
-        spyWinRating = 0;
-      } else if (exUser.totalCount - exUser.spyPlayCount === 0) {
-        voteSpyRating = 0;
-      }
-
-      spyWinRating =
-        (exUser.spyWinCount / exUser.spyPlayCount).toFixed(2) * 100;
-      voteSpyRating =
-        (
-          exUser.voteSpyCount /
-          (exUser.totalCount - exUser.spyPlayCount)
-        ).toFixed(2) * 100;
-
-      console.log('spyWinRating', spyWinRating);
-      console.log('voteSpyRating', voteSpyRating);
-
-      return {
-        accessToken,
-        userId: exUser._id,
-        nickname: exUser.nickname,
-        profileImg: exUser.profileImg,
-        totayPlayCount: exUser.totalCount,
-        spyPlayCount: exUser.spyPlayCount,
-        ctzPlayCount: exUser.totalCount - exUser.spyPlayCount,
-        spyWinRating,
-        voteSpyRating,
-      };
-    } else return;
-  };
-
-  createUserToken = async (kakaoUserInfo) => {
-    // DB에 유저 정보 없음 => 회원가입 / 토큰발급 / 토큰리턴
-    console.log('여기는 user-provider.js 2, createUserToken');
-
-    let nickNum, nickname, _id;
-    let allUser = await User.find();
-
-    // DB에 유저가 하나도 없다면 초기값 세팅
-    if (allUser.length === 0) {
-      _id = 1;
-      nickname = 'Agent_001';
-    } else {
-      // DB에 유저가 있을 경우
-      const lastNum = allUser.slice(-1)[0].nickname; // 마지막 document 의 nickname
-      let n = +lastNum.slice(6) + 1; // nickname 에서 Agent_ 뒷부분만 가져온 후 Number 변환
-      // n이 1000이상이면 Agent_ 뒤에 그대로 붙이고, 1000보다 작으면 001 의 형태로 붙이기
-      if (n < 1000) {
-        nickNum = (0.001 * n).toFixed(3).toString().slice(2);
-        nickname = `Agent_${nickNum}`;
-      } else {
-        nickname = `Agent_${n}`;
-      }
-      nickname = nickname;
-    }
-    // 위에서 만든 값으로 newUser DB 에 저장하기
-    const newUser = await User.create({
-      _id: +nickNum,
-      nickname,
-      email: kakaoUserInfo.kakao_account.email,
-      profileImg: kakaoUserInfo.properties.thumbnail_image
-        ? kakaoUserInfo.properties.thumbnail_image
-        : 'default',
-    });
-
-    console.log('여기는 user-provider.js 3, newUser::::::', newUser);
-
-    // 새로 생셩한 newUser에게 _id 값으로 토큰 발급
-    const newUserToken = await jwtService.createAccessToken(newUser._id);
-
-    // 클라이언트에 전달하기 위해 유저 정보 가공
-    let spyWinRating, voteSpyRating, totalCount;
-    if (newUser.totalCount === 0) {
-      spyWinRating = 0;
-      voteSpyRating = 0;
-    } else if (newUser.spyPlayCount === 0) {
-      spyWinRating = 0;
-    } else if (newUser.totalCount - newUser.spyPlayCount === 0) {
-      voteSpyRating = 0;
-    }
-
-    spyWinRating =
-      (newUser.spyWinCount / newUser.spyPlayCount).toFixed(2) * 100;
-    voteSpyRating =
-      (
-        newUser.voteSpyCount /
-        (newUser.totalCount - newUser.spyPlayCount)
-      ).toFixed(2) * 100;
-
-    // 전부 0일 때 자꾸 NaN 으로 나온다.
-    console.log('spyWinRating', spyWinRating);
-    console.log('voteSpyRating', voteSpyRating);
-
-    // 최종적으로 클라이언트에 넘겨주는 정보
-    return {
-      accessToken: newUserToken,
-      userId: newUser._id,
-      nickname: newUser.nickname,
-      profileImg: newUser.profileImg,
-      totayPlayCount: newUser.totalCount,
-      spyPlayCount: newUser.spyPlayCount,
-      ctzPlayCount: newUser.totalCount - newUser.spyPlayCount,
-      spyWinRating,
-      voteSpyRating,
-    };
-  };
-
   getUserInfo = async (decodedId, accessToken) => {
     const exUser = await User.findOne({
       _id: decodedId,
@@ -215,7 +79,7 @@ class UserProvider {
       },
       data: qs.stringify({
         grant_type: 'authorization_code',
-        client_id: process.env.CLIENT_ID,
+        client_id: process.env.CLIENT_ID_BACK,
         redirectUri: process.env.CALLBACK_URL_LOCAL,
         code: req.query.code,
       }),
@@ -227,6 +91,8 @@ class UserProvider {
     );
     res.header('Access-Control-Allow-Origin', '*');
     res.header('Authorization', kakaoToken.data.access_token);
+    res.header('Content-Type', 'text/html; charset=utf-8');
+
     return res.send({ accessToken: kakaoToken.data.access_token });
   };
 
@@ -240,6 +106,7 @@ class UserProvider {
     const { authorization } = req.headers;
     const [authType, kakaoToken] = (authorization || '').split(' ');
 
+    const bodyToken = req.body;
     const kakaoResult = await axios({
       method: 'get',
       url: 'https://kapi.kakao.com/v2/user/me',
@@ -252,10 +119,10 @@ class UserProvider {
     console.log(
       '------------여기는 user-provider.js의 getKakaoUserInfo---------------'
     );
-    console.log('kakaoToken:::', kakaoToken.data);
+    console.log('kakaoToken:::', kakaoToken);
     console.log('kakaoUserData:::', kakaoUserData);
 
-    const exUserInfo = await this.exUserGetToken(kakaoUserData);
+    const exUserInfo = await UserRefo.exUserGetToken(kakaoUserData);
 
     // 1. 가입한 유저 => 토큰 + 유저정보 바로 전달
     if (exUserInfo) {
@@ -263,10 +130,28 @@ class UserProvider {
       return res.status(200).json(exUserInfo);
     }
     // 2. 미가입 유저 => 회원가입 + 토큰발급 후 토큰 + 유저정보 전달
-    const newUserInfo = await this.createUserToken(kakaoUserData);
+    const newUserInfo = await UserRefo.createUserToken(kakaoUserData);
     console.log('user-route.js 5, newUserToken::::::', newUserInfo);
     return res.status(201).json(newUserInfo);
   };
 }
 
 module.exports = new UserProvider();
+
+/* 예시
+  user-route.js 1, res.locals에 저장한 userInfo:::::: {
+    id: 2519073484,
+    connected_at: '2022-11-10T01:02:16Z',
+    properties: { nickname: '미뇽' },
+    kakao_account: {
+      profile_nickname_needs_agreement: false,
+      profile_image_needs_agreement: true,
+      profile: { nickname: '미뇽' },
+      has_email: true,
+      email_needs_agreement: false,
+      is_email_valid: true,
+      is_email_verified: true,
+      email: 'alsuddl25@naver.com'
+    }
+  }
+*/
