+++ src/users/user-route.js
@@ -5,7 +5,6 @@ const router = express.Router();
 const loginMiddleware = require('../middlewares/login-middleware');
 const kakaoMiddleware = require('../middlewares/kakao-middleware');
 const UserProvider = require('./user-provider');
-const { RolesAnywhere } = require('aws-sdk');
 
 // 카카오 로그인 1 : 클라이언트에서 인가코드 전달 받음 =>  카카오로 토큰 요청 =>  클라이언트에 카카오 토큰 전달
 router.get('/api/auth/kakao/callback', UserProvider.getKakaoToken);
