+++ src/users/util.js
@@ -1,0 +1,7 @@
+const { User } = require('../schemas/user');
+const jwtService = require('./jwt');
+require('dotenv').config();
+
+class Utils {}
+
+module.exports = new Utils();
